pyproject.toml
[tool.poetry]
name = "locadoraVeiculos"
version = "0.1.0"
description = ""
authors = ["Arla Figueiredo <arla.figueiredo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
